backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0

# Database dependencies
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File processing
PyPDF2==3.0.1
python-docx==1.1.0
Pillow>=10.0.0

# RAG and ML dependencies
langchain==0.0.354
langchain-community==0.0.10
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
huggingface-hub>=0.19.4
transformers>=4.36.2
torch>=2.2.0

# HTTP requests
httpx>=0.25.2
aiofiles>=23.2.1

# Data processing
pandas>=2.1.4
numpy>=1.24.4

# Logging
python-json-logger>=3.3.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1